Remove stack-let
Symbolics stack-let does not have the same semantics as a let with
dynamic-extent declarations, so there is no easy way to define the
stack-let macro in a useful way.  It might be necessary to introduce
dynamic-extent declarations later, but the places where this is feasible
will have to be manually checked.